# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sorolens API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Sorolens indexes and surfaces data from Soroban smart contracts: events,
    invocations, storage, and per-contract analytics. This spec is the single
    source of truth for the auto-generated Go client (packages/go-client) and
    matches the live routes in apps/api/internal/router/router.go.

    Two namespaces are documented: the original `/api/v1` surface and the
    consistent `/api/v2` surface introduced in issue #144. v2 is additive; v1 is
    unchanged. See docs/api-v2.md for the field-by-field v1 -> v2 mapping.

    Authentication: most read endpoints accept an optional API key via
    `Authorization: Bearer <token>` or `X-API-Key: <token>`. API-key management
    additionally requires an admin identity via role headers
    (`X-User-ID`, `X-GitHub-Login`, `X-User-Sub`, or `X-GitHub-ID`).
  contact:
    name: Sorolens
servers:
  - url: https://api.sorolens.xyz
    description: Production
...
paths:
...
  /api/v1/rules:
    get:
      tags: [Alert rules]
      summary: List alert rules
      description: |
        Returns every stored rule, newest first. A rule is the source text of
        the Sorolens rule language plus its scope (contract, network) and the
        parsed `for` window.
      operationId: listAlertRules
      security:
        - ApiKeyAuth: [read:contracts]
      responses:
        "200":
          description: Rules, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: "#/components/schemas/AlertRule"
        "500":
          description: Internal error
    post:
      tags: [Alert rules]
      summary: Create an alert rule
      description: |
        Parses and validates `source` before storing it, so a malformed rule
        can never reach the evaluator. Returns 422 with a structured
        diagnostic (line, column, hint) when the rule does not validate, and
        409 when an identical rule already exists.
      operationId: createAlertRule
      security:
        - ApiKeyAuth: [write:contracts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, source]
              properties:
                name: { type: string }
                source:
                  type: string
                  example: "fee_per_invocation > 0.5 XLM for 5m"
                severity:
                  type: string
                  enum: [Info, Warning, Critical]
                  default: Warning
                contract_id: { type: string }
                network: { type: string }
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "409":
          description: An identical rule already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Rule did not validate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
...
  /api/v1/rules/library:
    get:
      tags: [Alert rules]
      summary: Curated sample rules
      description: |
        A small library of example rules that are guaranteed to validate.
        The dashboard offers them as starting points in the editor.
      operationId: listRuleLibrary
      security:
        - ApiKeyAuth: [read:contracts]
      responses:
        "200":
          description: Sample rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        description: { type: string }
                        severity: { type: string }
                        source: { type: string }
  /api/v1/rules/{id}:
    patch:
      tags: [Alert rules]
      summary: Enable or pause a rule
      operationId: setAlertRuleEnabled
      security:
        - ApiKeyAuth: [write:contracts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "404":
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Alert rules]
      summary: Delete a rule
      operationId: deleteAlertRule
      security:
        - ApiKeyAuth: [write:contracts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204":
          description: Rule deleted
        "404":
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

                description: See /api/v1/contracts/{id}/snapshot; already v2-conformant.
        "404":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "Scoped API key. Send as `Authorization: Bearer <token>` (or `X-API-Key`)."
    RoleAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: |
        Role-based auth. Resolve a user identity from one of
        X-User-ID, X-GitHub-Login, X-User-Sub, or X-GitHub-ID and enforce the
        minimum role declared next to this security scheme.
    IdentityHeaderAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: Watchlist endpoints read the user ID from X-User-ID.
...
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Global per-IP rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    AlertRule:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        source:
          type: string
          description: Rule text in the Sorolens rule language.
          example: "error_rate > 5% for 15m"
        severity: { type: string, enum: [Info, Warning, Critical] }
        contract_id: { type: string }
        network: { type: string }
        window: { type: string, example: "15m" }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: object
              required: [code, message]
              properties:
                code:
                  type: string
                  enum: [NOT_FOUND, INVALID_INPUT, INTERNAL, RATE_LIMITED, UNSUPPORTED_MEDIA_TYPE]
                message:
                  type: string
                request_id:
                  type: string
            - type: object
              properties:
                error:
                  type: string
...